feat(parser): new pipeline parser, all values can be indexed/accessed, etc.
Pipeline expressions cannot be followed by other expressions naked. It
needs to be enclosed in parentheses to use the value. This is because
only one callable is allowed in each step of the pipeline, except the
initial value, which can be any expression/identifier.

```
hi aqua = 1 + 2 | abs()~       >_< ok
hi aqua = 1 + 2 | abs() + 3~   >_< NOT OK
hi aqua = (1 + 2 | abs()) + 3~ >_< ok
```

Besides the two mentioned above:
- declarations no longer require data type.
- expressions can be a standalone statement as long as it is ended by a
terminator (`~`)