feat(parser): Add an expect, ok or recover, then finish pipeline for error recovery
This way, i can expect the next token, add an action to return any type
from a parser method if the expectation is ok, or recover using a
strategy is the expectation failed. Danced with generics...it was hard
but worth it.

feat(parser): derive Default for all production structs, implement for
enums

refactor(lexer): correct outdated doc comment