feat(parser): implement rest of statements error recovery
statement parsers now expect the starting tok in peek, to be consistent
with expression statement parsers which expect the initial value to be
in peek.

This means i can remove the `Parser::reverse` method since all statement
parsers now start in peek, not current.

feat(parser): add an expect peek is in method that does not advance
cursor when expectation is ok. This is just for `Parser::parse_body`
which expects a statement token in peek but should not advance it since
statement parsers expect the starting tok to be in peek.